% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEdges.R
\name{bamEdges}
\alias{bamEdges}
\title{Get edges of BAM alignment reads}
\usage{
bamEdges(bam.path)
}
\arguments{
\item{bam.path}{Path to the BAM file. The BAM file is presumed to contain
read PAIRs.}
}
\value{
A \code{GRanges} object representing edges of BAM reads.Usage cases
\itemize{
\item For ATAC sequencing, read edges are roughly transposition sites of the Tn5
transposase. However, to get the exact base-pair-precision
transposition sites you need to manually shift the read start/end positions.
\item Other cases?
}
}
\description{
\code{bamEdges()} reads a BAM file and returns the 5' AND 3' ends of EACH
paired-end read PAIR as a \code{GRanges} object.
}
