% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{breakGRanges}
\alias{breakGRanges}
\title{Break down a \code{GRanges} into a \code{GRangesList} of single ranges.}
\usage{
breakGRanges(gr)
}
\arguments{
\item{gr}{Input \code{GRanges}}
}
\value{
A \code{GRangesList} object
}
\description{
\code{breakGRanges()} breaks a \code{GRanges} object into a \code{GRangesList}
where each list element is a single range (i.e., row)
in the original \code{GRanges}.
}
