% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{consensusGRanges}
\alias{consensusGRanges}
\title{Compute 'consensus' GRanges from a list of GRanges objects}
\usage{
consensusGRanges(grs)
}
\arguments{
\item{grs}{A list of GRanges objects, could be length 1 or more.}
}
\value{
A GRanges object consisting the consensus ranges.
}
\description{
\code{consensusGRanges()} computes consensus GRanges from a list of
GRanges objects. Definition of 'consensus' is explained in Details section.
}
\section{Consensus algorithm TODO}{
The following steps are adopted to compute a single \code{GRanges} object from
a list of input \code{GRanges}. Inter-range methods of the \code{GenomicRanges} package
are used for computation.
\enumerate{
\item Concatenate all ranges from all GRanges in the list into a single GRanges.
\item Get a collection of non-overlapping ranges by \code{disjoin}.
\item Only retain non-overlapping ranges that overlaps with at least two ranges
from the original concatenated GRanges.
\item Merge connecting (i.e., distance is zero bp) ranges by \code{reduce}.
\item Start again the concatenated GRanges as in step 1, get ranges that
do not overlap with any other ranges by \code{countOverlaps}.
\item Concatenate ranges from step 4 and 5 to yield the final 'consensus'.
}
}

\seealso{
The following inter-range methods from the \code{GenomicRanges} package are used:
\itemize{
\item \link[GenomicRanges:inter-range-methods]{GenomicRanges::disjoin}
\item \link[GenomicRanges:inter-range-methods]{GenomicRanges::reduce}
\item \link[GenomicRanges:findOverlaps-methods]{GenomicRanges::countOverlaps}
}

For \code{countOverlaps}, this functions always adopt parameter \code{type = "any"}.
}
