% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEdges.R
\name{countEdges}
\alias{countEdges}
\title{Count number of BAM alignment read edges for each range of a \code{GRanges} object}
\usage{
countEdges(ranges, bam.path)
}
\arguments{
\item{ranges}{A GRanges object containing the ranges to consider}

\item{bam.path}{Path to the BAM file.}
}
\value{
The input GRanges object with an appended metadata column
\code{edge.counts}
}
\description{
\code{countEdges()} counts number of BAM read pair edges for each
range of a given \code{GRanges} object.
}
\seealso{
For more detailed discussion on "edges"
of a BAM alignment file, see \code{\link[=bamEdges]{bamEdges()}}.
Consult \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} about \code{GRanges} objects.
}
