% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEdges.R
\name{countEdgesCached}
\alias{countEdgesCached}
\title{Count number of BAM alignment read edges for each range of a \code{GRanges/GRangesList} object}
\usage{
countEdgesCached(ranges, edges)
}
\arguments{
\item{ranges}{A GRanges/GRangesList object containing the ranges to consider}

\item{edges}{A GRanges containing the edges. Use \code{bamEdges()} to get this.}
}
\value{
The input GRanges/GRangesList object with an appended metadata column
\code{edge.counts}.
}
\description{
\code{countEdgesCached()} counts number of BAM read pair edges for each
range of a given \code{GRanges} object. Compared to the exported \code{countEdges()},
\code{countEdgesCached()} requires providing the pre-extracted edges using \code{bemEdges()}.
}
\seealso{
For more detailed discussion on "edges"
of a BAM alignment file, see \code{\link[=bamEdges]{bamEdges()}}.
Consult \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} about \code{GRanges} objects.
}
