% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeCounterOptions.R
\name{ecOptions.default}
\alias{ecOptions.default}
\title{Defaults of global options}
\usage{
ecOptions.default()
}
\value{
A named list of the default values for the global options.
}
\description{
\code{ecOptions.default()} is internally used for global option
defaults
}
