% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeCounterOptions.R
\name{ecOptions}
\alias{ecOptions}
\title{Global options for the edgeCounter package}
\usage{
ecOptions(...)
}
\arguments{
\item{...}{Parameter name-value pairs supported by the package. See details
for a list of global parameters. If no parameter is provided, do not change
current options.}
}
\value{
Up-to-date edgeCounter options.
}
\description{
\code{ecOptions()} sets edgeCounter package global options.
}
\details{
Global options are accessible by \code{getOption("ecOptions")}.
}
\section{Supported options}{
\itemize{
\item \code{parallel}: a Boolean specifying whether to enable parallel computation.
\item \code{PSOCK.names}: PSOCK cluster \code{names} parameter.
See \link[parallel:makeCluster]{parallel::makePSOCKcluster}. Not used if \code{parallel} is false.
\item \code{parallel.packages}: Which packages to load in PSOCK cluster nodes.
}
\subsection{Parallel computation}{

This package always use PSOCK cluster as supported by the \code{parallel} package
for parallel computation. A 'diligent' approach for initialization
and clean-up of the PSOCK cluster is used, meaning that the PSOCK cluster is
initialized only when functions supporting parallel computation is called,
and the cluster is stopped whenever the called function exits.
}
}

\examples{
ecOptions() # Returns current option
ecOptions(parallel = FALSE) # Turns off parallel computation
}
