% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeCounterOptions.R
\name{ecOptions.valid}
\alias{ecOptions.valid}
\title{Validation of global options}
\usage{
ecOptions.valid()
}
\value{
A named list where names are names of available global options and
values are functions for validating the option.
Validation functions must take in a single argument and return Boolean.
}
\description{
\code{ecOptions.valid()} is internally used for global option
parameters sanity validation.
}
