% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecParallel.R
\name{ecParallel}
\alias{ecParallel}
\title{Set parallel computing options of edgeCounter}
\usage{
ecParallel(PSOCK.names)
}
\arguments{
\item{PSOCK.names}{\code{name} of \link[parallel:makeCluster]{parallel::makePSOCKcluster}.
If \code{NULL} is provided, parallel computing feature will be turned off.}
}
\value{
None.
}
\description{
\code{ecParallel()} sets global parallel computing options.
}
\examples{
ecParallel(6) # 6 workers, local machine
getOption("ecOptions")
ecParallel(NULL) # turn off PSOCK parallelization
getOption("ecOptions")
}
