% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExperiment.R
\name{edgeExperiment}
\alias{edgeExperiment}
\title{Construct a RangedSummarizedExperiment with BAM paths and ranges of interest}
\usage{
edgeExperiment(bam.paths, ranges)
}
\arguments{
\item{bam.paths}{A named character vector containing paths to the BAM files.
Names are used as sample IDs.}

\item{ranges}{A GRanges or GRangesList object containing the ranges for counting.
If GRanges is provided, each single range row is calculated individually.
If GRangesList is provided, each GRanges in the list is calculated individually
instead.}
}
\value{
A RangedSummarizedExperiment object containing the counting results.
One \code{assay} is saved, \code{counts}, containing the counting results.
\code{colData} will contain the sample IDs as row names but no column.
}
\description{
\code{edgeExperiment()} constructs a RangedSummarizedExperiment object containing
the edge-counting results given BAM file paths and a GRanges/GRangesList with
the target ranges for counting. You need to specify IDs of the samples.
}
\seealso{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}
}
