% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExperiment.R
\name{edgeExperimentFromCounts}
\alias{edgeExperimentFromCounts}
\title{Construct a RangedSummarizedExperiment with edge-counting results.}
\usage{
edgeExperimentFromCounts(edges, ranges)
}
\arguments{
\item{edges}{Pre-extracted edges using \code{\link[=edgeExtract]{edgeExtract()}}.}

\item{ranges}{A GRanges or GRangesList object containing the ranges for counting.
If GRanges is provided, each single range row is calculated individually.
If GRangesList is provided, each GRanges in the list is calculated individually
instead.}
}
\value{
A RangedSummarizedExperiment object containing the counting results.
One \code{assay} is saved, \code{counts}, containing the counting results.
\code{colData} will contain the sample IDs as row names but no column.
}
\description{
\code{edgeExperimentFromCounts()} constructs a RangedSummarizedExperiment
object containing the edge-counting results, given pre-extracted edges of
the samples and a GRanges/GRangesList with the target ranges for counting.
}
\section{Details}{
To extract edges of the samples use the \code{edgeExtract} function. This function
supports parallel computing as set up by \code{\link[=ecParallel]{ecParallel()}}.
}

\seealso{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment},
\code{\link[=edgeExtract]{edgeExtract()}}, \code{\link[=ecParallel]{ecParallel()}}.
}
