% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExperiment.R
\name{export.edgeExperiment}
\alias{export.edgeExperiment}
\title{Export edge-counting results}
\usage{
export.edgeExperiment(edgeExp, base.path, norm = "epm_ir")
}
\arguments{
\item{edgeExp}{A SummarizedExperiment object with counts assay data.
May be created by \code{\link[=edgeExperiment]{edgeExperiment()}} functions.}

\item{base.path}{Directory to save the exported files. Note that for
each sample one BED file will be exported as \verb{\{sampleID\}.bed}.}

\item{norm}{Normalization method. Currently supported are
\code{epm_ir}, \code{epm} and \code{none}. Refer to Details for explanation.}
}
\value{
None. Only side effect of saving BED files.
}
\description{
\code{export.edgeExperiment()} exports the edge-counting results
as BED tracks for visualization. Counts data normalization method may be
specified to allow comparison among different samples.
Current implementation is relatively slow due to \code{IRanges::RleList}
performance issues.
}
\section{Details}{
Nature of the \code{edgeExperiment()} function family is to generate counts
data for downstream analysis. Apart from differential analysis, which is
supported by various packages such as \code{edgeR} and \code{DESeq2}, visualization
is important for getting intuition of the data and verifying the results.

To support this, \code{export.edgeExperiment()} exports counts data into BED
tracks. Further, to allow comparison among different samples,
normalization methods as described below may be adopted.
\subsection{Normalization of counts}{

Problem: different samples have different numbers of reads. Solution:
calculate 'edges per million in range (\code{epm_ir})'. It divides all counts
by the total number of edges in the sample.

This function also allows normalization against ALL edges in a sample
with the \code{epm} option. Refer to \code{\link[=edgeExperimentFromCounts]{edgeExperimentFromCounts()}} for
more details.

No normalization: \code{none}.
}

\subsection{BED columns}{

Export follows standard 12-column BED format:

chrom, chromStart, chromEnd, name, score, strand
thickStart, thickEnd, itemRgb, blockCount, blockSizes, blockStarts

name is \code{names} of the \code{rowRanges}
score is (optionally normalized) count values with one decimal place
strand is always "." (i.e., no strand)
itemRgb is always 255,0,0 (i.e., red color)
blockCount/Sizes/Starts will be set according to the ranges
}

\subsection{BED file track line}{

One track line is written to first line of the BED file to
hopefully set up appropriate display settings. Refer to the
\href{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}{UCSC BED} and
\href{https://groups.google.com/g/igv-help/c/IyBzO5_j5E8}{IGV forum}.

As of November 2022, we only found \code{igv.js} as a browser that supports all
features.
}
}

\examples{
\dontrun{
export.edgeExperiment(exp, ".", norm = "epm_ir")
}
}
