% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBEDtrackLine}
\alias{getBEDtrackLine}
\title{Construct UCSC-format BED track line with score display}
\usage{
getBEDtrackLine(name, score)
}
\arguments{
\item{name}{Name of the track}

\item{score}{Scores of the track, for setting up score display limits.
Although there is no upper limit, scores MUST be >= 0.}
}
\value{
A string of "track line" without newline character.
}
\description{
\code{getBEDtrackLine()} constructs "track line" based on
\href{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}{UCSC guide}.
It tries to turn on score display based on settings suggested in
\href{https://groups.google.com/g/igv-help/c/IyBzO5_j5E8}{IGV help form}.
}
