% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importBED}
\alias{importBED}
\title{Reading BED and BED-like file with minimum dependency}
\usage{
importBED(path, extra.cols = NULL)
}
\arguments{
\item{path}{Path to the BED file. See details for format of the BED file.}

\item{extra.cols}{A named list for information on extra columns. It should
contain two elements \code{names} and \code{types} for name and type of the columns.}
}
\value{
A GRanges object
}
\description{
\code{importBED()} read in a BED file as a GRanges object.
}
\section{Details}{
\subsection{BED file format}{

The standard BED format supported by this function is three-column. Refer to
\href{https://genome.ucsc.edu/FAQ/FAQformat.html}{UCSC FAQ} for more details.
}

\subsection{Extra columns}{

Apart from the standard BED format, you can supply arbitrary extra columns by
providing names and types of the columns in the \code{extra.cols} optional argument.

The extra column names must be a character vector.

Types must follow the shorthand types used by \code{\link[readr:read_delim]{readr::read_delim()}}.
}

\subsection{Strand column}{

According to UCSC documentation, \code{strand} is an optional column at position 6.
Therefore, 6th column of the file, if existent, must specify strand info.
Supported strand spec characters are \code{.} (i.e., no strand), \code{+} and \code{-}.
}
}

