% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecParallel.R
\name{is.list.like}
\alias{is.list.like}
\title{Verify if an object is list-like}
\usage{
is.list.like(obj)
}
\arguments{
\item{obj}{A single object}
}
\value{
Boolean where TRUE means the object is 'list-like'.
}
\description{
\code{is.list.like()} checks if an object is 'list-like'.
}
\details{
A 'list-like' object is anything that has the following operators:
\itemize{
\item \verb{[[]]} double bracket subsetting, which should return an object in the list.
Integer subsetting must be supported.
\item \code{length} which should return length of the list.
}
}
\examples{
gr <- makeGRanges(
   tibble::tribble(
   ~start, ~end, ~seqnames,
   1, 100, "chr2L",
   80, 120, "chr2L",
   100, 300, "chrX"
))

is.list.like(gr)  # FALSE
is.list.like(GenomicRanges::GRangesList(list(gr, gr)))  # TRUE

}
