% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{readNarrowPeaks}
\alias{readNarrowPeaks}
\title{Read in MACS2 narrowPeaks from one/multiple files into GRanges objects}
\usage{
readNarrowPeaks(paths, names = NULL)
}
\arguments{
\item{paths}{A character vector of MACS2 narrowPeak file paths}

\item{names}{(optional) A character vector for names of the GRanges}
}
\value{
A named list of GRanges objects. Names are the input paths.
If \code{names} is provided, \code{names} is used
as list names instead of input paths.
}
\description{
\code{readNarrowPeaks()} reads in MACS2 narrowPeaks from files
into a list of GRanges object(s).
}
\examples{
np.path <- system.file("extdata", "readNarrowPeaks.example.narrowPeak",
    package = "edgeCounter"
)
readNarrowPeaks(np.path, names = "sample") # Returns a list of GRanges
}
