% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowRangesToBED}
\alias{rowRangesToBED}
\title{Convert rowRanges to BED-format}
\usage{
rowRangesToBED(edgeExp)
}
\arguments{
\item{edgeExp}{A \code{RangedSummarizedExperiment} object}
}
\value{
A list of BED-formatted rowRanges. See details.
}
\description{
\code{rowRangesToBED()} takes in a \code{RangedSummarizedExperiment}
and converts its \code{rowRanges()} GRangesList into BED-formatted list.
}
\section{Details}{
Returned list is essentially a data.frame with the following columns.
\code{x=y} below means that \code{x} is the list element in the returned list,
and \code{y} is the \code{GRanges} equivalent.

\if{html}{\out{<div class="sourceCode">}}\preformatted{chrom = seqnames; Each GRanges MUST contain only one unique seqnames.
chromStart = min(start())
chromEnd = max(end())
name = name of the GRanges defined in rowRanges list
strand = Always set to '.'
thickStart = min(start())
thickEnd = max(end())
itemRgb = Always set to '255,0,0'
blockCount = length()
blockSizes = width()  # comma-separated list
blockStarts = start() - min(start())  # comma-separated list
}\if{html}{\out{</div>}}
\subsection{Performance note}{

Current implementation is relatively slow due to the \code{vapply}s
involving \code{IRanges::RleList}.
}
}

