% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExperiment.R
\name{varMeanPlot}
\alias{varMeanPlot}
\title{Plots variance ~ mean of range counts}
\usage{
varMeanPlot(edgeExp, name = NULL)
}
\arguments{
\item{edgeExp}{Edge-counting results as constructed by \code{\link[=edgeExperiment]{edgeExperiment()}}.}

\item{name}{Optional. Extra identifier of the dataset to be added on title.}
}
\value{
None. Side effect of generating a single plot.
}
\description{
\code{varMeanPlot()} plots variance ~ mean of range counts given
edge-counting result \code{SummarizedExperiment} object.
}
\seealso{
\code{\link[=edgeExperiment]{edgeExperiment()}}, \code{\link[=edgeExperimentFromCounts]{edgeExperimentFromCounts()}}.
}
