% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExperiment.R
\name{edgeExperimentFromCounts}
\alias{edgeExperimentFromCounts}
\title{Construct a RangedSummarizedExperiment with edge-counting results.}
\usage{
edgeExperimentFromCounts(edges, ranges)
}
\arguments{
\item{edges}{Pre-extracted edges using \code{\link[=edgeExtract]{edgeExtract()}}.}

\item{ranges}{A GRanges or GRangesList containing the ranges for counting.
If GRanges is provided, each single range row is calculated individually.
If GRangesList is provided, each GRanges in the list
is calculated individually instead.}
}
\value{
A RangedSummarizedExperiment containing the counting results.
One \code{assay} is saved, \code{counts}, containing the counting results.
\code{colData} will contain the sample IDs as row names but no column.
}
\description{
\code{edgeExperimentFromCounts()} constructs a
RangedSummarizedExperiment object containing the edge-counting results,
given pre-extracted edges of the samples
and a GRanges/GRangesList with the target ranges for counting.
}
\section{Details}{
To extract edges of the samples use the \code{edgeExtract} function.
This function supports parallel computing as set up by \code{\link[=ecParallel]{ecParallel()}}.
\subsection{colData}{

sampleID = IDs provided as names of the \code{edges} input list.
edges.ir = Number of edges within the given ranges.
edges.total = Number of edges in the original \code{edges} input.
}
}

\examples{
bam.paths <- system.file("extdata", "sample.near.Gapdh1.bam",
    package = "edgeCounter"
)
names(bam.paths) <- "sample"
ranges <- makeGRanges(
    tibble::tribble(
        ~start, ~end, ~seqnames,
        7791000, 7794000, "chr2R", # inclusive of Gapdh1 locus
        80, 120, "chr2L", # random range
        100, 300, "chrX" # random range
    )
)
edges <- edgeExtract(bam.paths) # Reads edges of one sample named "sample"
exp <- edgeExperimentFromCounts(edges, ranges) # You can use other ranges
}
\seealso{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment},
\code{\link[=edgeExtract]{edgeExtract()}}, \code{\link[=ecParallel]{ecParallel()}}.
}
