% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExperiment.R
\name{edgeExtract}
\alias{edgeExtract}
\title{Extract edges from BAM files}
\usage{
edgeExtract(bam.paths)
}
\arguments{
\item{bam.paths}{A named character vector with paths to the BAM files.
Names are used as sample IDs.}
}
\value{
A compressed GRangesList with edges from each of the BAM files.
IDs will be saved in \code{names()}.
}
\description{
\code{edgeExtract()} is a simple wrapper of \code{bamEdges}
to extract edges of the BAM reads from multiple BAM files each of which
represents one sample. This step is IO and memory intensive.
Do NOT support parallel computing.
}
\examples{
bam.paths <- system.file("extdata", "sample.near.Gapdh1.bam",
    package = "edgeCounter"
)
names(bam.paths) <- "sample"
edges <- edgeExtract(bam.paths) # Reads edges of one sample named "sample"
}
